requirements.txt
openai
python-dotenv
autogen-agentchat
autogen-ext[openai, azure, web-surfer]
autogenstudio
markdown-pdf
semantic-kernel
rich
fastmcp
opentelemetry-exporter-otlp-proto-grpc
azure-monitor-opentelemetry-exporter
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry
lxml
docx-mailmerge2